tools/lessonforge/pyproject.toml
[project]
name = "lessonforge"
version = "0.1.0"
description = "AI-powered lecture video production pipeline for RPD course"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "edge-tts>=7.2",
    "mutagen>=1.47",
    "rich>=13.0",
    "srt>=3.5",
    "Pillow>=10.0",
]

[project.optional-dependencies]
elevenlabs = ["elevenlabs>=2.0"]
recording = [
    "pynput>=1.7",
    "obsws-python>=1.7",
    "auto-editor>=29.8",
    "moviepy>=1.0",
]
youtube = ["google-api-python-client>=2.0", "google-auth-oauthlib>=1.0"]
dev = ["pytest>=8.0", "ruff>=0.5"]

[project.scripts]
lessonforge = "lessonforge.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lessonforge"]

[tool.ruff]
target-version = "py39"
line-length = 100
